+++ src/refactoring/pages/admin/components/index.ts
@@ -1,0 +1,2 @@
+export * from './CouponSection';
+export * from './ProductsAdminSection';
